se_open_data/config.rb - check the SEOD_CONFIG env var for default config path
This fixes a bug in various `converter` scripts to date, which were
using this method rather than (as in other seod commands)
SeOpenData.load_config.  The consequence was that the `converter`
scripts were always using `default.conf`!  Usually this didn't matter
as there were no differences which mattered, but sometimes it does!