se_open_data/csv/converter/limesurveycore.rb - switch to calling add_postcode_lat_long
Now call add_postcode_lat_long instead of add_postcode_lat_long2.
this means the former is now exclusively the externally used one, and
the inner one is ripe for refactoring and simplication.

Note to do this, two new parameters needed to be added. T

- to_schema: which schema to use
- use_ordinance_survey: whether or not to use the OS postcode geocodeer.

The defaults means that if omitted, the old behaviour is preserved.