test/geocoding/source-data/oxford-expected.csv - update expected data
To match the generated data - as the addresses have not changed but
the locations have.

The source and expected data comes from the current oxford
dataset. That's generated on the murmurations-od71 branch, which was
marginally different - mainly in that the text "GB" was added to the
Country Name field.

In theory though these locations shouldn't change...