lib/se_open_data/cli.rb - allow manual override of deploy destination
Allow callers (and `seod`) to set the deploy destination and
ownership, overriding that set in the config.

This is intended to facilitate purely user-land installs of open-data
which want to put the outputs somewhere locally to be possible without
needing to alter all the configuration files (which typically point to
remote servers).