lib/se_open_data/cli.rb - adjust the run_all command for current use
We want to allow this to supercede the explicit steps in the
tools/deploy/crontab script if the `open-data` project, and allow
custom parameters supplied via environment variables (as explained in
the updated documentation string).

We are assuming that all data directories in the `open-data` project
will be knowingly importing the next version, v2.3.0, and therefore it
can use this new `run_all` command in the crontab script.